@@ -1,10 +1,4 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
-import { PdfViewerPlugin } from '@syncfusion/ej2-vue-pdfviewer';
 
-Vue.use(PdfViewerPlugin);
-
-new Vue({
-  el: '#app',
-  render: h => h(App)
-})
+createApp(App).mount('#app')
